Progress bar and Event Issue
Made progress bar work with water used, and fixed a issue with singletimewaterevents

+++ Assets/scripts/GameManagement/DebugGameManagerTool.cs
@@ -9,7 +9,7 @@ public class DebugGameManagerTool : MonoBehaviour {
 	int id;
 
 	[SerializeField]
-	int amountOfWater;
+	float amountOfWater;
 
 
 	private void Update()
